requirements.txt
ttkbootstrap
pillow
flet>=0.24.0
